Refactoring - Card Script, minor adjustments to combat framework, created cardEngine
Refactored code so that the functionality of cards is stored within cardEngine instead of within each individual code, makes the code easier to maintain over the long term and slightly easier to understand. Added position as a input for positional effect as to make it able to take in multiple positions for more flexibility through an array in case multiple positions are preferred and / or have different effects.