Card stacking update (WIP)
Cards can stack now and mostly work however there are a lot of bugs still that need to be ironed out, a few identified problems include cards failing to change where their pointers are pointing to, and also cards failing to update the slots correctly. A refactoring may also be needed as the code has become messy once again as the logic becomes more complicated to facilitate card stacking and proper logic.