MAJOR card revamp in progress
Card's themselves are getting a major overhaul, with the goal of getting better drag and drop precision as well as integrating the slot system (and positions) more. Each card will now also allow more complex action on an individual basis as well now which will add complexity  and strategy to the game!

Current features are majorly a WIP and may or may not be compatible with the old card system