Hotfix, fixed menus, fixed positions, corrected variable names in controls to best practices
Menu was linked to new file locations and the variable names for menu_open were fixed and converted to the new snake_case best practice format, swapped contains for has in the card engine as contains is not the correct term for godot. Should be ready for move towards release branch.