Optimization pt 2
Cards will now be hidden and processes disabled if off screen, this allows for over 500 cards to  be spawned and 240+fps. The cost of this however is that for each card on screen there is slightly more overhead for the cards on screen since there is an extra node which signals when exiting/entering the screen. However overall the FPS should be much improved. This is still being tested but should lead to performance gains on lower end machines.