Added simple targeting system
Added the optional simple targeting system which removes some of the customizability of the targeting system however is completely optional. The old 'complex' targeting system can still be used and both are included in the template.

Any excess damage from hitting a evil shadow worm after it is killed will be transferred over to the blind assassin worm.