New card spawning system
Cards now are spawned with a new method inside of CardReg, using CardReg.get_card(card_script), cards are returned. Instead of using the old spawning methods this new method creates a card using the base card template, and then updates the stats and sprite. This helps standardize cards in case any changes are made to cards across the board making it easier to create cards as the number of card increases. Instead of holding links to scene, card registry holds links to the card's gd scripts.