Save encryption + redundancy + checksums
Currently I am expanding the save system to have some encryption connected to the unique id of the system it is played on preventing the moving of saves between players. Along side this each save comes with a checksum and redundant backup saves in case of corruption.

As players likely will have multiple computers, as well as possibly may want to share saves with other players, I will add a in game way to unbind saves eventually, however at the same time I'll raise a flag on first time loading said save that it is from another computer and that you should only use said save if you trust the source it came from.

This + the way I formatted the json file should lower the chances of arbitrary code execution on the side of users, and warn them if they decide to risk using someone elses save.

The checksums will also warn users if the save was modified, however if they wish to use a modified save, it will likely be used to disable achievements or something else of the sorts.

Preventing players from being able to modify their saves I think is a useless and futile task, however I think that encouraging honest play as well as improving the player's security is a doable task, and giving them the tools to circumvent it easily and accessibly is important as long as they're aware of the risk.